rename files to chunker_internal_* to avoid installing the headers
