+++ cpp/src/parquet/chunker_internal_hashtable.py
